fix(widgetmanager): the selection condition was added
When adding a widget to the environment of 1 renderwindow and n renderers, selectedState does not
work properly in mouse move event. I think It compares the pokedRenderer that is passed to the mouse
move event and the Renderer that WindgetManager has, and if they are different, it seems that there
is no need to handle picking.